[rename]: 아카이브 - 타입 이름 변경 및 폴더 이동
