+++ src/entities/archive/model/index.tsx
@@ -1,0 +1,1 @@
+export * from './type'
